chore(access-list): Add function to access touched slots (#287)
Since there is no way to access the internal map, this PR introduces a
`fn touched_slots`. To avoid conflicting with the semantics of `fn
access_list`, I named it `touched_slots`. The reason for this is that
using the map speeds up the process of checking whether certain slots
have been touched.