fix: js tracer behavior (#295)
Closes https://github.com/paradigmxyz/revm-inspectors/issues/294 while
retaining fix for https://github.com/paradigmxyz/reth/issues/13089

The idea is that in steps (aka logs) we want `contract.address` to
always be the address of contract which storage is accessed
(`target_address`). However, when dealing with delegatecall frames,
`contract.address` should be set to the `bytecode_address` as it is
marking the destination of delegatecall.

This means we can't just forward `active_call().contract` to steps
anymore and need to re-construct the `Contract` value manually