feat(js): add logic to count current opcode cost instead of cumulative (#336)
Currently the revm-inspectors js tracer acts differently from the geth
implementation where the `getCost()` function returns the cumulative
amount of cost for each step instead of the individual opcode cost. I
added a simple function that will store the previous spent value and
then use that in the next step to calculate the currect opcode cost.